backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-socketio==5.3.6
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyjwt==2.9.0
requests==2.32.3
eventlet==0.36.1
